+++ ubo_app/constants.py
@@ -69,6 +69,7 @@ CONFIG_PATH = platformdirs.user_config_path(appname='ubo', ensure_exists=True)
 SECRETS_PATH = CONFIG_PATH / '.secrets.env'
 PERSISTENT_STORE_PATH = CONFIG_PATH / 'state.json'
 
+CACHE_PATH = platformdirs.user_cache_path(appname='ubo', ensure_exists=True)
 DATA_PATH = platformdirs.user_data_path(appname='ubo', ensure_exists=True)
 
 DISPLAY_BAUDRATE = int(os.environ.get('UBO_DISPLAY_BAUDRATE', '60_000_000'))
@@ -94,6 +95,7 @@ CORE_SERVICE_IDS = [
     'rgb_ring',
     'rpi_connect',
     'sensors',
+    'speech_recognition',
     'speech_synthesis',
     'ssh',
     'users',