+++ ubo_app/services/050-vscode/login_page.py
@@ -6,8 +6,8 @@ import pathlib
 import re
 import subprocess
 
-from _constants import CODE_BINARY_PATH
 from commands import install_service
+from constants import CODE_BINARY_PATH
 from kivy.clock import mainthread
 from kivy.lang.builder import Builder
 from kivy.properties import NumericProperty, StringProperty